chore: move popup menu tracking out of `controls`
