--- src/trackingModules/controls/index.js
@@ -1,7 +1,0 @@
-import popupMenuTracking from './popupMenu';
-
-export default {
-  __depends__: [
-    popupMenuTracking
-  ]
-};
