+++ src/trackingModules/index.js
@@ -1,9 +1,11 @@
 import controlsTracking from './controls/popupMenu';
 import diagramTracking from './diagram';
+import popupMenuTracking from './popupMenu';
 
 export default {
   __depends__: [
     controlsTracking,
     diagramTracking
+    popupMenuTracking,
   ]
 };
